Move BitTrackerOperations to Analysis namespace
Relocated BitTrackerOperations from Mosa.Compiler.Framework.Stages to Mosa.Compiler.Framework.Analysis without changing its logic. Updated all references and using directives in affected files to use the new namespace.
